Refactoring done now issues -> (Entity/Draw not visible + GUI Button)

+++ src/main/java/ch/g_7/graphite/core/Timer.java
@@ -17,7 +17,7 @@ public class Timer {
 		return (int) (1 / deltaMillis * 1000);
 	}
 	
-	public double getDeltaMillis() {
+	public float getDeltaMillis() {
 		return deltaMillis;
 	}
